requirements.txt
ultralytics>=8.3.0
flask>=3.0.0
pillow>=8.0.0
gunicorn>=22.0.0